api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==6.1.0